update uboot pack script

+++ scripts/bootloader.sh
@@ -8,34 +8,34 @@ die() {
 }
 
 [ -s "./env.sh" ] || die "please run ./configure first."
+
 . ./env.sh
+
 O=$1
 if [ ! -z $O ] ; then
 	BOARD=$O
 fi
-#U=$TOPDIR/out/${TARGET_PRODUCT}/100MB
-U=$TOPDIR/u-boot-rt
-UBOOTORG=$TOPDIR/rt-pack/rtk/bpi-w2/bin/uboot-bpi-w2-1MB.bin
 
-mkdir -p $U
-TMP_FILE=${U}/${BOARD}.tmp
+U=$TOPDIR/SD/${TARGET_PRODUCT}/100MB
+if [ ! -d $U ]; then
+	mkdir -p $U
+fi
+
+TMP_FILE=/tmp/${BOARD}.tmp
 IMG_FILE=${U}/${BOARD}-2k.img
+UBOOT=$TOPDIR/rt-pack/rtk/${TARGET_PRODUCT}/bin/u-boot.bin
 
 (sudo dd if=/dev/zero of=${TMP_FILE} bs=1M count=1) >/dev/null 2>&1
 LOOP_DEV=`sudo losetup -f --show ${TMP_FILE}`
-
-
-UBOOT=$TOPDIR/u-boot-rt/u-boot.bin
-
-(sudo dd if=$UBOOT	of=${LOOP_DEV} bs=1k seek=40) >/dev/null 2>&1
-
+(sudo dd if=$UBOOT of=${LOOP_DEV} bs=1k seek=40) >/dev/null 2>&1
 sudo sync
-
 sudo losetup -d ${LOOP_DEV}
-
 (dd if=${TMP_FILE} of=${IMG_FILE} bs=1k skip=2 count=1022 status=noxfer) >/dev/null 2>&1
 
-rm -f ${IMG_FILE}.gz
+if [ -f ${IMG_FILE}.gz ]; then
+	rm -f ${IMG_FILE}.gz
+fi
+
 echo "gzip ${IMG_FILE}"
 gzip ${IMG_FILE}
-rm -f ${TMP_FILE}
+sudo rm -f ${TMP_FILE}
